move sample to own package to prevent R file errors
- because the library shared the same package it would show errors inside the sample that the fields in R are not constant
