+++ sample/src/se/emilsjolander/sprinkles/sample/ChooseTagActivity.java
@@ -1,11 +1,12 @@
-package se.emilsjolander.sprinkles;
+package se.emilsjolander.sprinkles.sample;
 
 import java.util.List;
 
+import se.emilsjolander.sprinkles.Query;
 import se.emilsjolander.sprinkles.Query.OnQueryResultHandler;
-import se.emilsjolander.sprinkles.models.Note;
-import se.emilsjolander.sprinkles.models.NoteTagLink;
-import se.emilsjolander.sprinkles.models.Tag;
+import se.emilsjolander.sprinkles.sample.models.Note;
+import se.emilsjolander.sprinkles.sample.models.NoteTagLink;
+import se.emilsjolander.sprinkles.sample.models.Tag;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
