+++ sample/src/se/emilsjolander/sprinkles/sample/CreateNoteActivity.java
@@ -1,10 +1,11 @@
-package se.emilsjolander.sprinkles;
+package se.emilsjolander.sprinkles.sample;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.Locale;
 
-import se.emilsjolander.sprinkles.models.Note;
+import se.emilsjolander.sprinkles.Query;
+import se.emilsjolander.sprinkles.sample.models.Note;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
