+++ sample/src/se/emilsjolander/sprinkles/sample/MyApplication.java
@@ -1,8 +1,10 @@
-package se.emilsjolander.sprinkles;
+package se.emilsjolander.sprinkles.sample;
 
-import se.emilsjolander.sprinkles.models.Note;
-import se.emilsjolander.sprinkles.models.NoteTagLink;
-import se.emilsjolander.sprinkles.models.Tag;
+import se.emilsjolander.sprinkles.Migration;
+import se.emilsjolander.sprinkles.Sprinkles;
+import se.emilsjolander.sprinkles.sample.models.Note;
+import se.emilsjolander.sprinkles.sample.models.NoteTagLink;
+import se.emilsjolander.sprinkles.sample.models.Tag;
 import android.app.Application;
 
 public class MyApplication extends Application {
